requirements/requirements-all.txt
colorama>=0.3.9
eido>=0.2.4
jinja2
logmuse>=0.2.0
pandas>=2.0.2
pephubclient>=0.4.0
pipestat>=0.12.0a1
peppy>=0.40.6
pyyaml>=3.12
rich>=9.10.0
ubiquerg>=0.8.1a1
yacman==0.9.3
pydantic-argparse>=0.9.0
psutil